Copy changes to start directory
