+++ start/inventory/src/main/java/io/openliberty/guides/inventory/InventoryManager.java
@@ -12,34 +12,46 @@
 //end::copyright[]
 package io.openliberty.guides.inventory;
 
+import java.util.ArrayList;
 import java.util.Properties;
 import io.openliberty.guides.inventory.client.SystemClient;
 import io.openliberty.guides.inventory.model.InventoryList;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
+import java.util.List;
+import java.util.Collections;
 
 import org.eclipse.microprofile.opentracing.Traced;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
 import io.opentracing.ActiveSpan;
 import io.opentracing.Tracer;
+import io.openliberty.guides.inventory.model.*;
 
 @ApplicationScoped
 public class InventoryManager {
 
-    private InventoryList invList = new InventoryList();
+    private List<SystemData> systems = Collections.synchronizedList(new ArrayList<>());
     private InventoryUtils invUtils = new InventoryUtils();
 
     public Properties get(String hostname) {
         Properties properties = invUtils.getProperties(hostname);
-        if (properties != null) {
-            invList.addToInventoryList(hostname, properties);
-        }
         return properties;
     }
 
+    public void add(String hostname, Properties systemProps) {
+        Properties props = new Properties();
+        props.setProperty("os.name", systemProps.getProperty("os.name"));
+        props.setProperty("user.name", systemProps.getProperty("user.name"));
+
+        SystemData system = new SystemData(hostname, props);
+        if (!systems.contains(system)) {
+                systems.add(system);
+        }
+    }
+
     public InventoryList list() {
-        return invList;
+        return new InventoryList(systems);
     }
 
 }
